Initial prototype commit
The result of many weeks of trying to get something like this working. Draws everything at 6x similarly to the hires-renderer branch of this mod, but in a way that doesn't cause input lag or dozens of rendering glitches (currently more like 6). The rendering pipeline is roughly as follows:
- Draw the background to a small (i.e. 320x180) buffer, then scale it up into a large (1920x1080) compositing buffer immediately, (i.e. way before the usual upscaling step)
- Draw the gameplay small and scale it up, but offset it based on the smoothed "camera" position, which for all other purposes is now zeroed out since we're doing that manually and it's the cause of the extremely glitchy stylegrounds
- Draw the lighting, bloom, etc, all scaled up manually (will need to convert these to IL hooks)
- Draw the foreground by adjusting its scale matrix field. It's impossible to render it to a buffer and nearest-neighbor upscale it like with the background, since that would render all the blending states useless (believe me, I've tried). This method works, but certain objects like foreground waterfalls and wind get drawn not snapped to the grid. This is a problem to fix later.
Ditch the 6x scaling normally applied and draw this composited buffer straight back to the screen, using the extend method to fill in the edges. I may add support for the zoom and border methods later.